+++ plugins/sanity-plugin-media/src/components/ReduxProvider/index.tsx
@@ -1,6 +1,6 @@
 import {type AnyAction, configureStore, type Store} from '@reduxjs/toolkit'
 import type {SanityClient} from '@sanity/client'
-import {Component, type ReactNode} from 'react'
+import {type ReactNode, useState} from 'react'
 import {Provider} from 'react-redux'
 import {createEpicMiddleware} from 'redux-observable'
 import type {AssetSourceComponentProps, SanityDocument} from 'sanity'
@@ -22,75 +22,72 @@ type Props = {
   selectedAssets?: AssetSourceComponentProps['selectedAssets']
 }
 
-class ReduxProvider extends Component<Props> {
-  store: Store
+function createReduxStore(props: Props): Store {
+  const epicMiddleware = createEpicMiddleware<AnyAction, AnyAction, RootReducerState>({
+    dependencies: {
+      client: props.client,
+    },
+  })
 
-  constructor(props: Props) {
-    super(props)
-
-    // Initialize redux store + middleware
-    const epicMiddleware = createEpicMiddleware<AnyAction, AnyAction, RootReducerState>({
-      dependencies: {
-        client: props.client, // inject sanity client as a dependency to all epics
+  const store = configureStore({
+    reducer: rootReducer,
+    middleware: (getDefaultMiddleware) =>
+      getDefaultMiddleware({
+        /*
+        serializableCheck: {
+          ignoredActions: [
+            assetsActions.deleteError.type,
+            uploadsActions.uploadRequest.type,
+            uploadsActions.uploadStart.type,
+          ]
+        },
+        */
+        // TODO: remove once we're no longer storing non-serializable data in the store
+        serializableCheck: false,
+        thunk: false,
+      }).prepend(epicMiddleware),
+    devTools: true,
+    preloadedState: {
+      assets: {
+        ...assetsInitialState,
+        assetTypes: isSupportedAssetType(props?.assetType) ? [props.assetType] : ['file', 'image'],
       },
-    })
-    this.store = configureStore({
-      reducer: rootReducer,
-      middleware: (getDefaultMiddleware) =>
-        getDefaultMiddleware({
-          /*
-          serializableCheck: {
-            ignoredActions: [
-              assetsActions.deleteError.type,
-              uploadsActions.uploadRequest.type,
-              uploadsActions.uploadStart.type,
-            ]
-          },
-          */
-          // TODO: remove once we're no longer storing non-serializable data in the store
-          serializableCheck: false,
-          thunk: false,
-        }).prepend(epicMiddleware),
-      devTools: true,
-      preloadedState: {
-        assets: {
-          ...assetsInitialState,
-          assetTypes: isSupportedAssetType(props?.assetType)
-            ? [props.assetType]
-            : ['file', 'image'],
-        },
-        debug: {
-          badConnection: false,
-          enabled: false,
-        },
-        dialog: {items: []},
-        notifications: {items: []},
-        search: {facets: [], query: ''},
-        selected: {
-          assets: props.selectedAssets || [],
-          document: props.document,
-          documentAssetIds: props.document ? getDocumentAssetIds(props.document) : [],
-        },
-        tags: {
-          allIds: [],
-          byIds: {},
-          creating: false,
-          fetchCount: -1,
-          fetching: false,
-          panelVisible: true,
-        },
-        uploads: {
-          allIds: [],
-          byIds: {},
-        },
+      debug: {
+        badConnection: false,
+        enabled: false,
       },
-    })
-    epicMiddleware.run(rootEpic)
-  }
+      dialog: {items: []},
+      notifications: {items: []},
+      search: {facets: [], query: ''},
+      selected: {
+        assets: props.selectedAssets || [],
+        document: props.document,
+        documentAssetIds: props.document ? getDocumentAssetIds(props.document) : [],
+      },
+      tags: {
+        allIds: [],
+        byIds: {},
+        creating: false,
+        fetchCount: -1,
+        fetching: false,
+        panelVisible: true,
+      },
+      uploads: {
+        allIds: [],
+        byIds: {},
+      },
+    },
+  })
 
-  override render() {
-    return <Provider store={this.store}>{this.props.children}</Provider>
-  }
+  epicMiddleware.run(rootEpic)
+
+  return store
+}
+
+function ReduxProvider({children, ...props}: Props) {
+  const [store] = useState(() => createReduxStore(props))
+
+  return <Provider store={store}>{children}</Provider>
 }
 
 export default ReduxProvider